Architecture is derived from Streamlet, not Implementation
