+++ crates/til_vhdl/src/ir/implementation.rs
@@ -1,14 +1,11 @@
 use tydi_common::{
-    error::{Result, TryOptional, TryResult},
+    error::{Result, TryResult},
     name::{Name, NameSelf},
 };
 
-use tydi_vhdl::{
-    architecture::{arch_storage::Arch, Architecture},
-    declaration::Declare,
-};
 
-use super::{context::Context, IntoVhdl, Ir};
+
+use super::{context::Context};
 
 #[derive(Debug, Clone, PartialEq, Eq, Hash)]
 pub struct Implementation {
@@ -44,64 +41,10 @@ impl Implementation {
     pub fn kind(&self) -> &ImplementationKind {
         &self.kind
     }
-
-    pub fn resolve_for(
-        &self,
-        ir_db: &dyn Ir,
-        arch_db: &mut dyn Arch,
-        prefix: impl TryOptional<Name>,
-    ) -> Result<String> {
-        match self.kind() {
-            ImplementationKind::Structural(context) => {
-                let arch_body = context.canonical(ir_db, arch_db, prefix)?;
-                let mut architecture = Architecture::from_database(arch_db, "Behavioral")?;
-                architecture.add_body(arch_db, &arch_body)?;
-
-                let result_string = architecture.declare(arch_db)?;
-                arch_db.set_architecture(architecture);
-
-                Ok(result_string)
-            }
-            ImplementationKind::Link => todo!(),
-        }
-    }
 }
 
 impl NameSelf for Implementation {
     fn name(&self) -> &Name {
         &self.name
-    }
-}
-
-impl IntoVhdl<String> for Option<Implementation> {
-    fn canonical(
-        &self,
-        ir_db: &dyn Ir,
-        arch_db: &mut dyn Arch,
-        prefix: impl TryOptional<Name>,
-    ) -> Result<String> {
-        match self {
-            Some(implementation) => match implementation.kind() {
-                ImplementationKind::Structural(context) => {
-                    let arch_body = context.canonical(ir_db, arch_db, prefix)?;
-                    let mut architecture = Architecture::from_database(arch_db, "Behavioral")?;
-                    architecture.add_body(arch_db, &arch_body)?;
-
-                    let result_string = architecture.declare(arch_db)?;
-                    arch_db.set_architecture(architecture);
-
-                    Ok(result_string)
-                }
-                ImplementationKind::Link => todo!(),
-            },
-            None => {
-                let architecture = Architecture::from_database(arch_db, "Behavioral")?;
-
-                let result_string = architecture.declare(arch_db)?;
-                arch_db.set_architecture(architecture);
-
-                Ok(result_string)
-            }
-        }
     }
 }
